Split out int and float properties to separate files
