+++ glass-renderer/property_float.h
@@ -1,0 +1,9 @@
+#ifndef PROPERTY_FLOAT
+#define PROPERTY_FLOAT
+
+#include "property.h"
+
+extern PropertyTypeHandler property_float;
+
+
+#endif
