+++ glass-renderer/property_int.h
@@ -1,0 +1,8 @@
+#ifndef PROPERTY_INT
+#define PROPERTY_INT
+
+#include "property.h"
+
+extern PropertyTypeHandler property_int;
+
+#endif
